test polars transformer
