+++ requirements.txt
@@ -1,5 +1,6 @@
 numpy
-pandas
+pandas>=1.5.3
+polars==1.5.0
 scikit-learn
 torch
 uvicorn
